Skip re-installing Fish if version is as expected

+++ ubuntu-install.sh
@@ -7,11 +7,17 @@ set -euxo pipefail
 # https://launchpad.net/~fish-shell/+archive/ubuntu/release-3/+packages
 # (because Ubuntu 20.04 LTS packages an ancient Fish v3.1.0 which is 1.5+ years behind current >v3.3+)
 
-sudo apt-get remove -y fish fish-common
+FISH_VERSION=$(fish --version | cut -c 15-)
+if $(dpkg --compare-versions $FISH_VERSION lt 3.5.1); then
+    echo "Fish version is <3.5.1, so upgrading it..."
+    sudo apt-get remove -y fish fish-common
 
-sudo apt-add-repository -y ppa:fish-shell/release-3 
-sudo apt-get update -y && sudo apt-get upgrade -y
-sudo apt-get install -y fish
-fish --version
+    sudo apt-add-repository -y ppa:fish-shell/release-3
+    sudo apt-get update -y && sudo apt-get upgrade -y
+    sudo apt-get install -y fish
+    fish --version
+else
+    echo "Fish version is >=3.5.1, great!"
+fi
 
 ./apt-install.sh
